Move lib/loginuid to session/loginuid
