+++ lib/srv/reexec.go
@@ -46,7 +46,6 @@ import (
 
 	"github.com/gravitational/teleport"
 	"github.com/gravitational/teleport/api/types"
-	"github.com/gravitational/teleport/lib/loginuid"
 	"github.com/gravitational/teleport/lib/pam"
 	"github.com/gravitational/teleport/lib/selinux"
 	"github.com/gravitational/teleport/lib/service/servicecfg"
@@ -61,6 +60,7 @@ import (
 	logutils "github.com/gravitational/teleport/lib/utils/log"
 	"github.com/gravitational/teleport/lib/utils/uds"
 	"github.com/gravitational/teleport/session/auditd"
+	"github.com/gravitational/teleport/session/loginuid"
 )
 
 // FileFD is a file descriptor passed down from a parent process when
